requirements.txt
Django
djangorestframework
pandas
